Add files via upload
The text you want to display isn't the only parameter you can use with display.scroll.  You can control the speed at which the text scrolls across the display by adding a an additional parameter to the display.scroll method called delay. The larger the value for the delay, the slower the text will scroll