Create Remember and Use Values
A variable can hold different kinds of values, such as integers (whole numbers),  decimal numbers, and characters.  Each of these kinds of values uses a different data type. The declaration a = 42 creates a variable named a.  The value 42 is an integer, so the Python language automatically assigns the int data type to the a variable, and gives an initial value of 42. If you do not want an initial value, you can assign a value of None to the variable, such as a = None.
Next, c = "m" declares a variable named c and initializes it to the value "m." Since "m" is a character,  Python assigns c the data type str for storing strings or text.    
Then, d = 42.06  declares a variable named d.  The value 42.06 is a decimal number, so Python assigns d the data type float, which can hold floating-point decimal values.