Create Simple Decisions
The program decisions.py demonstrates single-operation decision making.  It compares the value of a to b, and sends a message to tell you whether or not a is greater than b, with an if…else statement.  

If the condition (a > b) is True, it executes the if statement’s code block: display.scroll("a is greater than b").  If it so happens a is actually not greater than b, it executes the else code block instead: display.scroll("a is not greater than b").