Create Count and Repeat
A for loop is typically used to make the statements in a code block repeat a certain number of times.  For example, your cyber:bot may use five different values to make a sensor detect distance, so it needs to repeat a certain code block five times.  For this task, we use a for loop.  Here is an example that uses a for loop to count from 1 to 10 and display the values.