Create Types of Comments
First, it has a block comment at the top.  Technically this block comment is called a docstring. Docstrings start with a triple quote """ and end with a triple quote """. People often use these to create block comments and you can write as many lines of notes in between a pair of triple quotes as you want. Use caution when placing docstrings in your code as they could become the documentation that is associated with a function or method.

Also, each line of code has a line comment starting with # to its right, explaining what the code does. Line comments can also be placed above a section of code if they are very long, to avoid having to scroll sideways to read it.

In Python, constants don’t exist at the level of coding we are currently describing, however, we can still use styling to indicate to other people that may read our code that we are intending certain variables to behave as constants, meaning we don’t expect these variables to actually change during an execution of the program.