Create get_presses.py
To the left and right side of the LED display, the micro:bit module has two tactile pushbuttons labeled A and B. The micro:bit module can be programmed to respond to these buttons being pressed. Each button exists as an object referred to as button_a and button_b.