Create was_pressed.py
Another commonly used method of the buttons is to be able to tell if the button was pressed since the micro:bit module turned on, or since the last time the code checked to see if it was pressed,  whichever is most recent. The button_a.was_pressed method returns True or False depending on whether or not button A was pressed.