Create store_change_retrieve_global.py
If you want to permanently change the global variable from within the function, you must use the keyword global to call upon the global variable. In store_change_retrieve.py, k is created as a global variable with the value 5 assigned. Each time k is called upon in the function add_k it gets incremented by 1